src/backend/requirements.txt
flask>=3.0,<4.0
gunicorn>=22.0,<23.0
msal>=1.28,<2.0
msgraph-sdk>=1.5,<2.0
azure-ai-contentunderstanding>=1.0.1,<2.0
azure-ai-projects>=2.0.0,<3.0
azure-cosmos>=4.7,<5.0
azure-identity>=1.17,<2.0
